Fix for signal checkpoint writing unconverged soln
This commit patches the issue (#25755) where a checkpoint can write an
unconverged solution depending on when the USR1 signal is received, thereby
affecting the accuracy of recovered simulations. The fix 1) adds logic that
ensures a checkpoint is not output unless _current_execute_flag is contained
in _execute_on, and 2) enforces that Checkpoint’s ‘execute_on’ parameter may
only be set to ‘TIMESTEP_END’ so that only converged solutions are output.
Exodiff tests are added to ensure that simulations recovered from signal-based
checkpoints result in the same solutions as the uninterrupted simulations.