Testing with shell=False.
This commit tests all tests without launching them in a separate shell.
A custom string splitting utility function was implemented to correctly split
the commands into lists to be passed to subprocess.Popen. If we procees with
this approach, this utility will need to be cleaned up for conciseness and
clarity. Some of the tests files were changes to reflect the fact that the
tests are no longer run in a separate shell by default.